+++ CHANGELOG.md
@@ -22,6 +22,7 @@ v 2019.4 (?) Changes in this release:
 - Mark resource as failed when code loading fails (#1520)
 - Print extra env variables in init log and only store those in database (#1482)
 - Add get_environment_id to plugin context (#1331)
+- Log server bind address and bind port on startup
 
 IMPORTANT CHANGES:
 - Older compiler versions are no longer supported with this server